lib: Introduce INV count in the fanout decision-making
Count how many INVs have been received and how many peers we have fanout to.
Use that to keep fanning out or start reconciling

+++ hyper-lib/src/lib.rs
@@ -73,7 +73,7 @@ pub struct OutputResult {
     unreachable_count: usize,
     in_poisson_mean: u16,
     out_poisson_mean: u16,
-    out_fanout: usize,
+    out_fanout: u64,
     in_fanout: f32,
     n: u32,
     erlay: bool,
